LOGBROKER-10215 Fix
