+++ ydb/core/persqueue/public/cloud_events/ya.make
@@ -3,6 +3,7 @@ LIBRARY()
 SRCS(
     actor.cpp
     actor.h
+    cloud_events.h
     events_writer.cpp
     events_writer.h
 )
@@ -10,8 +11,10 @@ SRCS(
 PEERDIR(
     contrib/libs/protobuf
     library/cpp/json
+    library/cpp/yt/coding
     library/cpp/unified_agent_client
     library/cpp/monlib/dynamic_counters
+    ydb/core/persqueue/events
     ydb/core/persqueue/public/cloud_events/proto
     ydb/core/protos
     ydb/core/protos/schemeshard
