+++ ydb/core/protos/pqconfig.proto
@@ -225,8 +225,10 @@ message TPQConfig {
 
     message TCloudEventsConfig {
         optional bool Enabled = 1 [default = false];
-        optional string UaURI = 2;
-        optional string FilePath = 3;
+        oneof Endpoint {
+            string UaURI = 2;
+            string FilePath = 3;
+        }
     }
 
     optional TCloudEventsConfig CloudEventsConfig = 61;
