+++ ydb/core/tx/schemeshard/ya.make
@@ -372,6 +372,7 @@ PEERDIR(
     ydb/core/persqueue/public
     ydb/core/persqueue/public/partition_index_generator
     ydb/core/persqueue/public/partition_key_range
+    ydb/core/persqueue/public/cloud_events
     ydb/core/persqueue/writer
     ydb/core/protos
     ydb/core/resource_pools
